# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SceneIt one-video ingestion and semantic-search proof.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /proof/searches/{searchId}/frames/{rank}:
    parameters:
      - { name: searchId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: rank, in: path, required: true, schema: { type: integer, minimum: 1, maximum: 5 } }
    get:
      operationId: getProofFrame
      tags: [proof]
      summary: Read a retained source frame for a shared proof result
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: JPEG source frame
          content:
            image/jpeg:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/ProtectedFailure" }
        "403": { $ref: "#/components/responses/ProtectedFailure" }
        "404": { $ref: "#/components/responses/ProtectedFailure" }
...
  /proof/source:
    get:
      operationId: streamProofSource
      tags: [proof]
      summary: Stream the rights-approved original with HTTP byte-range support
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Complete original source
          content:
            video/mp4:
              schema: { type: string, format: binary }
        "206":
          description: Requested byte range from the original source
          content:
            video/mp4:
              schema: { type: string, format: binary }
        "404":
          $ref: "#/components/responses/Failure"
        "416":
          $ref: "#/components/responses/ProtectedFailure"
        "401": { $ref: "#/components/responses/ProtectedFailure" }
        "403": { $ref: "#/components/responses/ProtectedFailure" }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: __Host-sceneit_session }
    csrfToken: { type: apiKey, in: header, name: X-CSRF-Token }
    stripeSignature: { type: apiKey, in: header, name: Stripe-Signature }
...
  responses:
...
    Failure:
      description: Safe, actionable failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
    ProtectedFailure:
      description: Typed protected-resource failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProtectedApiFailure"
...
  schemas:
...
    ApiFailure:
      type: object
      required: [error, code]
      properties:
        error: { type: string }
        code: { type: string }
    ProtectedApiFailure:
      type: object
      additionalProperties: false
      required: [error, code, state]
      properties:
        error: { type: string }
        code: { type: string }
        state:
          type: string
          enum: [admission_required, quota_exhausted, processing, uncertain, service_unavailable, not_found, unauthorized]
        retryable: { type: boolean, default: false }
        retryAfterSeconds: { type: ["integer", "null"], minimum: 0 }
...